+++ pkg/tenantbillingserver/stripe/webhooks.tti.go
@@ -12,6 +12,7 @@ import (
 	"github.com/stripe/stripe-go/webhook"
 	"go.thethings.network/lorawan-stack/v3/pkg/events"
 	"go.thethings.network/lorawan-stack/v3/pkg/log"
+	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 )
 
 const (
@@ -84,8 +85,10 @@ func (s *Stripe) handleWebhook(c echo.Context) error {
 		return s.createTenant(ctx, sub)
 	case stripe.SubscriptionStatusIncomplete, stripe.SubscriptionStatusIncompleteExpired:
 		return nil
-	case stripe.SubscriptionStatusCanceled, stripe.SubscriptionStatusPastDue, stripe.SubscriptionStatusUnpaid:
-		return s.suspendTenant(ctx, sub)
+	case stripe.SubscriptionStatusPastDue:
+		return s.updateTenantState(ctx, sub, ttnpb.STATE_FLAGGED)
+	case stripe.SubscriptionStatusCanceled, stripe.SubscriptionStatusUnpaid:
+		return s.updateTenantState(ctx, sub, ttnpb.STATE_SUSPENDED)
 	default:
 		logger.Errorf("Unhandled Stripe subscription status: %s", sub.Status)
 		return nil